Refatoracão da classe Game
